tehtävät 20.8

+++ Teht-2-Variables/2.2.py
@@ -1,5 +1,5 @@
 import math
 
-sade = int(input("Anna ympyrän säde: "))
+sade = float(input("Anna ympyrän säde: "))
 pinta_ala = math.pi * (sade ** 2)
 print(f"Ympyrän pinta-ala on: {pinta_ala}")
